ENH Implement \message
This uses a pure system for macro expansion with an impure loop around
it. It is sort main :: [Responses] -> [Request] IO model that Simon P
Jones mentions as the first IO model for Haskell.

It works pretty well. \input sort of works too, but it is not yet
correctly parsing the file names. I need to check exactly what TeX does
in similar situations.