ENH loadfont & selectfont
This is the full machinery for switching fonts. Now, all that is needed
is an implementation of the \font command (which will build on
loadfont/selectfont).

Also, instead of hardcoding the cmr10 load, it would be nice to have a
sort of TeX prefix that is prepended to the real input.