RFCT Move all Macro expansion into TkS Monad
This is also now implemented on top of a State Monad and with an added
environment variable.

This makes the code much simpler in cases like:

\def\a{...}
\if\a\b...\else\...

where before we had a bit of a hack. Also, we will be able to better
support some types of "inline expansion" without all these hacks (by
inline expansion, I mean reading expanded tokens while processing
something else, which is what TeX does).