TST Simple method to test expanded
This performs the same as putting files into tests/expanded.

However, it performs it in the same process as hextest, instead of
spawning a new hex process for each test. In fact, running the test
suite is starting to be a slowish operation and as more tests are added
it would only get worse.